add fees adapter for allbridge-core
